extract alchemist roll data
